refactor: extract stubFetch into @theholocron/http-client/testing
stubFetch was copy-pasted verbatim across 8 packages (clerk, confluence,
doppler, github, infisical, neon, postman, vercel, zendesk). The github
variant had a 3-line drift (extra TOKEN/REPO constants), showing the
expected decay from living in 8 places.

- Export stubFetch and FetchCall from a new src/testing.ts module in
  @theholocron/http-client, built to dist/testing.mjs via a new tsdown
  entry; exposed under the ./testing subpath export
- Collapse all eight helpers.ts copies to a single re-export line
- Keep github-client TOKEN and REPO constants in its own helpers.ts
- Switch all @theholocron/http-client deps to workspace:^ so tests
  resolve the local build (the ./testing subpath does not exist in the
  currently published version)
- Add outputs: [dist/**] to turbo build task so turbo caches and
  restores build artifacts; add ^build dependency to test task so
  http-client is built before any package's tests run
- Exclude src/testing.ts from http-client's own coverage (it is only
  exercised by the other packages' test suites)

Closes #160
