+++ packages/http-client/tsdown.config.ts
@@ -1,1 +1,6 @@
-export { default } from "@theholocron/tsdown-config/presets/library";
+import { library } from "@theholocron/tsdown-config/presets/library";
+
+export default library({
+	entry: ["src/index.ts", "src/testing.ts"],
+	deps: { neverBundle: [/^@theholocron\//, /^vitest/] },
+});
