+++ packages/postman-client/package.json
@@ -33,7 +33,7 @@
     "typecheck": "tsc --noEmit"
   },
   "dependencies": {
-    "@theholocron/http-client": "catalog:"
+    "@theholocron/http-client": "workspace:^"
   },
   "devDependencies": {
     "@types/node": "catalog:",
